which if apt-get present - install sudo only in that case.

+++ src/scripts/install_doppler.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 set +e
 
-apt-get install -y sudo
-ls -al /bin/sh && sudo rm /bin/sh && sudo ln -s /bin/bash /bin/sh && ls -al /bin/sh
+# check if apt-get is present
+if command -v apt-get &> /dev/null
+then
+  # install sudo
+  apt-get install -y sudo
+  ls -al /bin/sh && sudo rm /bin/sh && sudo ln -s /bin/bash /bin/sh && ls -al /bin/sh
+fi
 
 (curl -Ls --tlsv1.2 --proto "=https" --retry 3 https://cli.doppler.com/install.sh || wget -t 3 -qO- https://cli.doppler.com/install.sh) | sudo sh
 doppler --version
